sensors/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --extra-index-url=https://www.piwheels.org/simple requirements.in
#
--extra-index-url https://www.piwheels.org/simple

adafruit-blinka==8.38.0
    # via
    #   adafruit-circuitpython-busdevice
    #   adafruit-circuitpython-connectionmanager
    #   adafruit-circuitpython-dht
    #   adafruit-circuitpython-requests
    #   adafruit-circuitpython-typing
adafruit-circuitpython-busdevice==5.2.6
    # via adafruit-circuitpython-typing
adafruit-circuitpython-connectionmanager==1.0.1
    # via adafruit-circuitpython-requests
adafruit-circuitpython-dht==4.0.4
    # via -r requirements.in
adafruit-circuitpython-requests==3.2.2
    # via adafruit-circuitpython-typing
adafruit-circuitpython-typing==1.10.3
    # via
    #   adafruit-blinka
    #   adafruit-circuitpython-busdevice
adafruit-platformdetect==3.62.0
    # via adafruit-blinka
adafruit-pureio==1.1.11
    # via adafruit-blinka
ads1015==1.0.0
    # via automationhat
annotated-types==0.6.0
    # via pydantic
automationhat==0.4.1
    # via -r requirements.in
i2cdevice==1.0.0
    # via ads1015
numpy==1.26.4
    # via st7735
paho-mqtt==2.0.0
    # via -r requirements.in
pydantic==2.6.4
    # via pydantic-settings
pydantic-core==2.16.3
    # via pydantic
pydantic-settings==2.2.1
    # via -r requirements.in
pyftdi==0.55.0
    # via adafruit-blinka
pyserial==3.5
    # via pyftdi
python-dotenv==1.0.1
    # via pydantic-settings
pyusb==1.2.1
    # via pyftdi
rpi-gpio==0.7.1
    # via
    #   adafruit-blinka
    #   automationhat
rpi-ws281x==5.0.0
    # via adafruit-blinka
smbus==1.1.post2
    # via -r requirements.in
smbus2==0.4.3
    # via
    #   i2cdevice
    #   sn3218
sn3218==3.0.0
    # via automationhat
spidev==3.6
    # via st7735
st7735==1.0.0
    # via automationhat
sysv-ipc==1.1.0
    # via adafruit-blinka
typing-extensions==4.10.0
    # via
    #   adafruit-circuitpython-typing
    #   pydantic
    #   pydantic-core
